Restored boore_atkinson_2008.py
