+++ openquake/hazardlib/gsim/akkar_cagnan_2010.py
@@ -24,9 +24,9 @@ import numpy as np
 from scipy.constants import g
 
 from openquake.hazardlib.gsim.atkinson_boore_2006 import (
-    GMPE, BooreAtkinson2008, _get_site_amplification_linear,
-    _get_site_amplification_non_linear)
-from openquake.hazardlib.gsim.base import CoeffsTable
+     _get_site_amplification_linear, _get_site_amplification_non_linear)
+from openquake.hazardlib.gsim.boore_atkinson_2008 import BooreAtkinson2008
+from openquake.hazardlib.gsim.base import CoeffsTable, GMPE
 from openquake.hazardlib import const
 from openquake.hazardlib.imt import PGA, PGV, SA
