+++ openquake/hazardlib/gsim/can15/nbcc2015_aa13.py
@@ -13,8 +13,8 @@ from openquake.hazardlib import const
 from openquake.hazardlib.imt import PGA, PGV, SA
 from openquake.hazardlib.gsim.base import add_alias
 from openquake.hazardlib.gsim.atkinson_boore_2006 import (
-    BooreAtkinson2008, _get_site_amplification_non_linear,
-    _get_site_amplification_linear)
+    _get_site_amplification_non_linear, _get_site_amplification_linear)
+from openquake.hazardlib.gsim.boore_atkinson_2008 import BooreAtkinson2008
 
 dirname = os.path.dirname(__file__)
 BASE_PATH_AA13 = os.path.join(dirname, 'nbcc2015_tables')
